backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.23.1
pydantic>=2.10.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.2
requests>=2.31.0
python-multipart>=0.0.6
openai>=1.0.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
stripe>=8.0.0
